+++ types/upgrade.go
@@ -32,4 +32,10 @@ const (
 
 	// Veld is the upgrade name for Veld upgrade
 	Veld = "Veld"
+
+	// Mongolian is the upgrade name for Mongolian upgrade
+	Mongolian = "Mongolian"
+
+	// Altai is the upgrade name for Altai upgrade
+	Altai = "Altai"
 )