+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## V1.9.1
+This release introduces the Altai upgrade
+
+## V1.9.0
+This release introduces the Mongolian upgrade
+
 ## V1.8.0
 This release introduces the Veld upgrade